test: move unit test to specific folder
