+++ test/unit/coordinate.js
@@ -1,7 +1,6 @@
-/* global describe, it */
 import proj4 from 'proj4';
 import assert from 'assert';
-import Coordinates from '../src/Core/Geographic/Coordinates';
+import Coordinates from '../../src/Core/Geographic/Coordinates';
 
 // Define projection that we will use (taken from https://epsg.io/3946, Proj4js section)
 proj4.defs('EPSG:3946', '+proj=lcc +lat_1=45.25 +lat_2=46.75 +lat_0=46 +lon_0=3 +x_0=1700000 +y_0=5200000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs');
