+++ test/unit/featureUtils.js
@@ -1,10 +1,9 @@
-import GeoJsonParser from '../src/Parser/GeoJsonParser';
-import FeaturesUtils from '../src/Renderer/ThreeExtended/FeaturesUtils';
-import Coordinates from '../src/Core/Geographic/Coordinates';
-/* global describe, it */
+import assert from 'assert';
+import GeoJsonParser from '../../src/Parser/GeoJsonParser';
+import FeaturesUtils from '../../src/Renderer/ThreeExtended/FeaturesUtils';
+import Coordinates from '../../src/Core/Geographic/Coordinates';
 
-const assert = require('assert');
-const geojson = require('./data/geojson/simple.geojson.json');
+const geojson = require('../data/geojson/simple.geojson.json');
 
 const promise = GeoJsonParser.parse(geojson, { crsOut: 'EPSG:4326', buildExtent: true });
