+++ test/unit/geojson.js
@@ -1,10 +1,9 @@
 import proj4 from 'proj4';
-import GeoJsonParser from '../src/Parser/GeoJsonParser';
-/* global describe, it */
+import assert from 'assert';
+import GeoJsonParser from '../../src/Parser/GeoJsonParser';
 
-const assert = require('assert');
-const holes = require('./data/geojson/holes.geojson.json');
-const gpx = require('./data/geojson/gpx.geojson.json');
+const holes = require('../data/geojson/holes.geojson.json');
+const gpx = require('../data/geojson/gpx.geojson.json');
 
 proj4.defs('EPSG:3946',
     '+proj=lcc +lat_1=45.25 +lat_2=46.75 +lat_0=46 +lon_0=3 +x_0=1700000 +y_0=5200000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs');
