+++ test/unit/obb.js
@@ -1,12 +1,11 @@
 import * as THREE from 'three';
 import proj4 from 'proj4';
 import assert from 'assert';
-import Extent from '../src/Core/Geographic/Extent';
-import BuilderEllipsoidTile from '../src/Core/Prefab/Globe/BuilderEllipsoidTile';
-import PlanarTileBuilder from '../src/Core/Prefab/Planar/PlanarTileBuilder';
-import TileGeometry from '../src/Core/TileGeometry';
-import OBB from '../src/Renderer/ThreeExtended/OBB';
-/* global describe, it */
+import Extent from '../../src/Core/Geographic/Extent';
+import BuilderEllipsoidTile from '../../src/Core/Prefab/Globe/BuilderEllipsoidTile';
+import PlanarTileBuilder from '../../src/Core/Prefab/Planar/PlanarTileBuilder';
+import TileGeometry from '../../src/Core/TileGeometry';
+import OBB from '../../src/Renderer/ThreeExtended/OBB';
 
 const max = new THREE.Vector3(10, 10, 10);
 const min = new THREE.Vector3(-10, -10, -10);
