+++ test/unit/provider_url.js
@@ -1,8 +1,7 @@
-/* global describe, it */
 /* eslint-disable no-template-curly-in-string */
 import assert from 'assert';
-import URLBuilder from '../src/Provider/URLBuilder';
-import Extent from '../src/Core/Geographic/Extent';
+import URLBuilder from '../../src/Provider/URLBuilder';
+import Extent from '../../src/Core/Geographic/Extent';
 
 const layer = {};
