+++ test/unit/scheduler.js
@@ -1,6 +1,5 @@
-/* global describe, it */
 import assert from 'assert';
-import Scheduler from '../src/Core/Scheduler/Scheduler';
+import Scheduler from '../../src/Core/Scheduler/Scheduler';
 
 const scheduler = new Scheduler();
 global.window = {
